Added ability to read e820 entries

+++ src/kernel/string.c
@@ -1,6 +1,6 @@
 #include "string.h"
 
-void num_to_str (u64 num, char *str) {
+void num_to_str (u32 num, char *str) {
   int i = 0;
   do {
     str[i++] = num % 10 + '0';
